requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.2.0
backcall==0.2.0
black==20.8b1
click==7.1.2
coverage==5.3
decorator==4.4.2
Flask==1.1.2
gazpacho==1.1
iniconfig==1.1.1
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.9.0
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
python-dotenv==0.14.0
regex==2020.10.15
selenium==3.141.0
six==1.15.0
toml==0.10.1
tqdm==4.50.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
